
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIM4rv.R
\name{cauchyp}
\alias{cauchyp}
\title{Cauchy combination test (Cauchy‐p)}
\usage{
cauchyp(x)
}
\arguments{
\item{x}{a numeric vector containing p-values}
}
\value{
a combined p-value indicating the joint effect
}
\description{
This function combines p-values using Cauchy combination test
for the omnibus test of testing the joint genetic effect.
}
