% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ex2_covar.R
\docType{data}
\name{Ex2_covar}
\alias{Ex2_covar}
\title{An example dataset of covariate file}
\format{
An object of class \code{"data.frame"} listing IDs and covariates separately
}
\usage{
data(Ex2_covar)
}
\description{
Small, artificially generated toy data set that
provides artificial information of covariates for 15 individuals
to illustrate the pre-processing with the use of the package.
}
\examples{

data(Ex2_covar)
head(Ex2_covar)

}
\references{
This data set was artificially created and modified for the ZIM4rv package.
}
\keyword{datasets}
