% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ex2_dosage.R
\docType{data}
\name{Ex2_dosage}
\alias{Ex2_dosage}
\title{An example dataset of dosage file}
\format{
An object of .dosage file
}
\usage{
data(Ex2_dosage)
}
\description{
Small, artificially generated toy data set that
provides artificial information of dosage for 15 individuals
to illustrate the pre-processing with the use of the package.
}
\examples{

data(Ex2_dosage)
head(Ex2_dosage)

}
\references{
This data set was artificially created and modified for the ZIM4rv package.
}
\keyword{datasets}
