% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ex2_fam.R
\docType{data}
\name{Ex2_fam}
\alias{Ex2_fam}
\title{An example dataset of .fam file}
\format{
An object of standard .fam file
}
\usage{
data(Ex2_fam)
}
\description{
Small, artificially generated toy data set that
provides artificial information of .fam file for 15 individuals
to illustrate the pre-processing with the use of the package.
}
\examples{

data(Ex2_fam)
head(Ex2_fam)

}
\references{
This data set was artificially created and modified for the ZIM4rv package.
}
\keyword{datasets}
