% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ex2_pheno.R
\docType{data}
\name{Ex2_pheno}
\alias{Ex2_pheno}
\title{An example dataset of pheno file}
\format{
An object of class \code{"data.frame"} listing IDs and phenotypes separately
}
\usage{
data(Ex2_pheno)
}
\description{
Small, artificially generated toy data set that
provides artificial information of phenotypes for 15 individuals
to illustrate the pre-processing with the use of the package.
}
\examples{

data(Ex2_pheno)
head(Ex2_pheno)

}
\references{
This data set was artificially created and modified for the ZIM4rv package.
}
\keyword{datasets}
