% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ex2_region.R
\docType{data}
\name{Ex2_region}
\alias{Ex2_region}
\title{An example dataset of genetic region file}
\format{
An object of class \code{"data.frame"} listing genetic regions
where each row contains chromosome, basepairs and the name of genetic region respectively
}
\usage{
data(Ex2_region)
}
\description{
Small, artificially generated toy data set that
provides artificial information of 3 genetic regions
to illustrate the pre-processing with the use of the package.
}
\examples{

data(Ex2_region)
head(Ex2_region)

}
\references{
This data set was artificially created and modified for the ZIM4rv package.
}
\keyword{datasets}
