
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIM4rv.R
\name{fisherp}
\alias{fisherp}
\title{Fisher's method (Fisher‐p)}
\usage{
fisherp(x)
}
\arguments{
\item{x}{a numeric vector containing p-values}
}
\value{
a combined p-value indicating the joint effect
}
\description{
This function combines p-values using Fisher's method
for the omnibus test of testing the joint genetic effect.
}
