% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zimfrv.R
\name{p_burden_single}
\alias{p_burden_single}
\title{p_burden_single}
\usage{
p_burden_single(wt, G_rare, s)
}
\arguments{
\item{wt}{a numeric vector containing weights for all variants}

\item{G_rare}{a data frame containing data of rare variants}

\item{s}{a matrix of the score statistics for each variant from each subject}
}
\value{
the p-value for the burden test
}
\description{
Compute the p-value for the burden test

This function takes a vector of weights, a data frame of rare variants
and a matrix of Score statistics produced by \code{U_fi_lmd} for ZIP model
or \code{U_phi_mu4zinb} for ZINB model to compute the p-value for the burden test.
}
