% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zimfrv.R
\name{p_kernel_single}
\alias{p_kernel_single}
\title{p_kernel_single}
\usage{
p_kernel_single(wt_matrix2, G_rare, s)
}
\arguments{
\item{wt_matrix2}{a diagonal matrix containing the squared weights for all variants}

\item{G_rare}{a data frame containing data of rare variants}

\item{s}{a matrix of the score statistics for each variant from each subject}
}
\value{
the p-value for the kernel test (ZIP-k)
}
\description{
Compute the p-value for the kernel test

This function takes a diagonal matrix of weights, a data frame of rare variants
and a matrix of Score statistics produced by \code{U_fi_lmd} for ZIP model
or \code{U_phi_mu4zinb} for ZINB model to compute the p-value for the kernel test.
}
