% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIM4rv.R
\name{phi_lambda_hat}
\alias{phi_lambda_hat}
\title{Estimation of phi_hat and lambda_hat for ZIP model}
\usage{
phi_lambda_hat(simud)
}
\arguments{
\item{simud}{a data frame containing a phenotype named y and covariates}
}
\value{
a list of 2 estimations of parameters for each subject
}
\description{
This function gives the estimation of 2 parameters phi and lambda for each subject
under the null hypothesis.
}
\details{
This function first fits zero‐inflated Poisson regression of phenotype y
on the covariates only to obtain the estimates of regression coefficients
and then compute the estimations of phi and lambda.
}
\seealso{
\code{zeroinfl}
}
