% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zimfrv.R
\name{preprocess_phenodata}
\alias{preprocess_phenodata}
\title{preprocess_phenodata}
\usage{
preprocess_phenodata(pheno_file, cov_file)
}
\arguments{
\item{pheno_file}{phenotype file in PLINK format}

\item{cov_file}{covariate file in PLINK format}
}
\value{
a data frame containing phenotypes and covariates respectively for all individuals in the required format for model fitting and testing
}
\description{
Preprocess phenotype files in PLINK format

This function converts PLINK format files into data frames containing phenotypes and covariates information in proper format
for the model fitting and testing.
}
\examples{
data(Ex2_pheno)
data(Ex2_covar)
preprocess_phenodata(Ex2_pheno,Ex2_covar)

}
