% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zimfrv.R
\name{U_fi_lmd}
\alias{U_fi_lmd}
\title{U_fi_lmd}
\usage{
U_fi_lmd(simudata, G_rare)
}
\arguments{
\item{simudata}{a data frame containing a phenotype named y and covariates}

\item{G_rare}{a data frame containing data of rare variants with the same subject order as in simudata}
}
\value{
a list of 2 matrice of the score statistics for each variant from each subject
}
\description{
Compute Score statistics for ZIP model

This function takes the estimations of phi and lambda produced by the \code{phi_lambda_hat}
and computes the score statistics under the null hypothesis.
}
