% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zimfrv.R
\name{U_phi_mu4zinb}
\alias{U_phi_mu4zinb}
\title{U_phi_mu4zinb}
\usage{
U_phi_mu4zinb(simudata, G_rare)
}
\arguments{
\item{simudata}{a data frame containing a phenotype named y and covariates}

\item{G_rare}{a data frame containing data of rare variants with the same subject order as in simudata}
}
\value{
a list of 2 matrice of the score statistics for each variant from each subject
}
\description{
Compute score statistics for ZINB model

This function takes the estimations of phi and lambda produced by the \code{phi_lambda_hat4negbin}
and computes the score statistics for ZINB model under the null hypothesis.
}
