% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vuong_test.R
\name{vuong_test}
\alias{vuong_test}
\title{vuong_test}
\usage{
vuong_test(phenodata)
}
\arguments{
\item{phenodata}{a data frame containing family and individual IDs for all objects as well as
zero-inflated counts as a phenotype and a set of covariates.
Each row represents a different individual.
The first two columns are Family ID (FID) and Individual ID (IID) respectively.
There must be one and only one phenotype in the third column and
the phenotype have to be zero-inflated count data which should be non-negative integers, e.g. neuritic plaque counts.
Each of the rest of columns represents a different covariate, e.g. age, sex, etc.}
}
\value{
nothing returned, prints a table of 3 test statistics and p values, and exits silently.
}
\description{
Vuong's test

This function performs Vuong's test, a likelihood ratio test for model selection and
non-nested hypotheses.
This function is for model selection between zero-inflated Poisson model and zero-inflated negative binomial model.
}
