% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zimfrv.R
\name{cauchyp}
\alias{cauchyp}
\title{cauchyp}
\usage{
cauchyp(x)
}
\arguments{
\item{x}{a numeric vector containing p-values}
}
\value{
a combined p-value indicating the joint effect
}
\description{
Cauchy combination test (Cauchy‐p)

This function combines p-values using Cauchy combination test
for testing the joint genetic effect.
}
